refactor: rename xccl files
