+++ paddle/fluid/distributed/collective/ProcessGroupBKCL.cc
@@ -15,7 +15,7 @@
 #include "paddle/fluid/distributed/collective/ProcessGroupBKCL.h"
 
 #include "paddle/fluid/distributed/collective/BKCLTools.h"
-#include "paddle/fluid/distributed/collective/Common.h"
+#include "paddle/fluid/distributed/collective/common.h"
 #include "paddle/fluid/platform/device/xpu/bkcl_helper.h"
 #include "paddle/fluid/platform/device/xpu/xpu_info.h"
 #include "paddle/fluid/platform/device_context.h"
