+++ paddle/fluid/distributed/collective/common.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "paddle/fluid/distributed/collective/Common.h"
+#include "paddle/fluid/distributed/collective/common.h"
 
 namespace paddle {
 namespace distributed {
