+++ paddle/fluid/distributed/collective/common.h
