+++ paddle/fluid/distributed/collective/process_group_mpi.cc
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "paddle/fluid/distributed/collective/ProcessGroupMPI.h"
+#include "paddle/fluid/distributed/collective/process_group_mpi.h"
 #include <chrono>
-#include "paddle/fluid/distributed/collective/Common.h"
+#include "paddle/fluid/distributed/collective/common.h"
 
 constexpr int64_t kWaitBlockTImeout = 10;
 namespace paddle {
