+++ paddle/fluid/pybind/CMakeLists.txt
@@ -155,21 +155,21 @@ if(WITH_CUSTOM_DEVICE)
 endif()
 
 if(WITH_PYTHON)
-  set(PYBIND_DEPS ${PYBIND_DEPS} processgroup eager_reducer)
+  set(PYBIND_DEPS ${PYBIND_DEPS} process_group eager_reducer)
   if(WITH_NCCL OR WITH_RCCL)
     set(PYBIND_DEPS ${PYBIND_DEPS} process_group_nccl)
   endif()
   if(WITH_XPU_BKCL)
-    set(PYBIND_DEPS ${PYBIND_DEPS} processgroup_bkcl)
+    set(PYBIND_DEPS ${PYBIND_DEPS} process_group_bkcl)
   endif()
   if(WITH_GLOO)
-    set(PYBIND_DEPS ${PYBIND_DEPS} processgroup_gloo)
+    set(PYBIND_DEPS ${PYBIND_DEPS} process_group_gloo)
   endif()
   if(WITH_MPI)
-    set(PYBIND_DEPS ${PYBIND_DEPS} processgroup_mpi)
+    set(PYBIND_DEPS ${PYBIND_DEPS} process_group_mpi)
   endif()
   if(WITH_CUSTOM_DEVICE)
-    set(PYBIND_DEPS ${PYBIND_DEPS} processgroup_custom)
+    set(PYBIND_DEPS ${PYBIND_DEPS} process_group_custom)
   endif()
   if(CMAKE_CXX_COMPILER_VERSION VERSION_GREATER 7.0)
     set(DISTRIBUTE_COMPILE_FLAGS "${DISTRIBUTE_COMPILE_FLAGS} -faligned-new")
