+++ paddle/phi/backends/CMakeLists.txt
@@ -65,12 +65,12 @@ if(WITH_CUSTOM_DEVICE)
     DEPS phi_capi)
 endif()
 
-set(COMM_UTILS_DEPS processgroup)
+set(COMM_UTILS_DEPS process_group)
 if(WITH_NCCL OR WITH_RCCL)
   set(COMM_UTILS_DEPS ${PROCESS_GROUP_UTILS_DEPS} process_group_nccl)
 endif()
 if(WITH_CUSTOM_DEVICE)
-  set(COMM_UTILS_DEPS ${PROCESS_GROUP_UTILS_DEPS} processgroup_custom)
+  set(COMM_UTILS_DEPS ${PROCESS_GROUP_UTILS_DEPS} process_group_custom)
 endif()
 cc_library(
   processgroup_comm_utils
