+++ paddle/phi/kernels/CMakeLists.txt
@@ -78,7 +78,7 @@ set(COMMON_KERNEL_DEPS
     gpc
     utf8proc)
 
-set(COMMON_KERNEL_DEPS ${COMMON_KERNEL_DEPS} processgroup)
+set(COMMON_KERNEL_DEPS ${COMMON_KERNEL_DEPS} process_group)
 if(WITH_NCCL OR WITH_RCCL)
   set(COMMON_KERNEL_DEPS ${COMMON_KERNEL_DEPS} process_group_nccl)
 endif()
